Adding new features
## Sitemap support

Setting a sitemap object on the .yaml allows to fetch urls from a given endpoint.
An optional `ignore` array can be provided to be able to exclude certain urls.

## Name tag

Allows for tags to be defined on the URL object so they can be repeated. Useful in conjunction with css to produce different page layouts.